requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
azure-core==1.31.0
azure-cosmos==4.7.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
portalocker==2.10.1
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.0
